adding pall operation

+++ test/print_error.c
@@ -1,9 +1,9 @@
 #include "monty.h"
 
-int print_error(int flag, size_t line_number)
+int print_error(int flag, unsigned int line_number)
 {
 	char *errors[] = {"usage: push integer", "nil"};
 
-	fprintf(stderr, "L%ld: %s\n", line_number, errors[flag]);
+	fprintf(stderr, "L%d: %s\n", line_number, errors[flag]);
 	return (EXIT_FAILURE);
 }
